requirements.txt
absl-py==2.1.0
accelerate==0.21.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
anyio==4.2.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bitsandbytes==0.41.0
bleach==6.1.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cmake==3.28.1
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
deepspeed==0.9.5
defusedxml==0.7.1
docker-pycreds==0.4.0
einops==0.6.1
einops-exts==0.0.4
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.109.0
fastjsonschema==2.19.1
ffmpy==0.3.1
filelock==3.13.1
flash-attn==2.4.2
flatbuffers==23.5.26
fonttools==4.47.2
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.12.2
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
gradio==3.35.2
gradio_client==0.2.9
grpcio==1.60.1
h11==0.14.0
h5py==3.10.0
hjson==3.1.0
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.20.3
idna==3.6
ipykernel==6.29.0
ipython==8.21.0
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.0.12
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
linkify-it-py==2.0.2
lit==17.0.6
-e git+https://github.com/haotian-liu/LLaVA.git@4afcebd2f668fa12900acbdb9402e46911721ad8#egg=llava
Markdown==3.5.2
markdown-it-py==2.2.0
markdown2==2.4.12
MarkupSafe==2.1.4
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.2.0
mpmath==1.3.0
multidict==6.0.4
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
notebook_shim==0.2.3
numpy==1.26.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.2.5.6
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.2.142
nvidia-cuda-nvcc-cu12==12.2.140
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.2.140
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.2.140
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.4.25
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.8.103
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.3.141
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.5.2.141
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.2.141
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.16.5
nvidia-nvjitlink-cu12==12.2.140
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opt-einsum==3.3.0
orjson==3.9.12
overrides==7.7.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.1
parso==0.8.3
peft==0.4.0
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==4.23.4
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.14
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2023.4
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.17.1
rsa==4.9
safetensors==0.4.2
scikit-learn==1.2.2
scipy==1.12.0
semantic-version==2.10.0
Send2Trash==1.8.2
sentencepiece==0.1.99
sentry-sdk==1.40.0
setproctitle==1.3.3
shortuuid==1.0.11
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
starlette==0.35.1
svgwrite==1.4.3
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0.post1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.35.0
tensorrt-bindings==8.6.1
tensorrt-libs==8.6.1
termcolor==2.4.0
terminado==0.18.0
threadpoolctl==3.2.0
timm==0.6.13
tinycss2==1.2.1
tokenizers==0.15.0
tomli==2.0.1
toolz==0.12.1
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.36.2
triton==2.0.0
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
uc-micro-py==1.0.2
uri-template==1.3.0
urllib3==2.2.0
uvicorn==0.27.0.post1
wandb==0.16.2
wavedrom==2.0.3.post3
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
wrapt==1.14.1
yarl==1.9.4
